{"schema":"decodex/commit/1","summary":"Split MCP project control modules","authority":"XY-1115"}
